+++ client/src/components/Custom/CustomContent/CanvasComponent.tsx
