+++ client/src/share/ModalComponentCustom.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import Modal from 'react-modal';
 import styled from 'styled-components';
-import modal_cart from '../assets/images/img_modal/modal_cart.png';
-import modal_cake from '../assets/images/img_modal/modal_cake.png';
-import CustomSidebar from '../components/custom/CustomSidebar';
-import CustomContent from '../components/custom/CustomContent';
+import CustomSidebar from '../components/Custom/CustomContent/CustomSidebar';
+import CustomContent from '../components/Custom/CustomContent/CustomContent';
+import ModalButtons from '../components/Custom/CustomButton/ModalButtons';
 type ModalProps = {
   isOpen: boolean;
   onRequestClose: () => void;
@@ -56,12 +55,6 @@ const ModalContainer = styled.div`
   flex-direction: column;
 `;
 
-const ModalButtons = styled.div`
-  align-self: flex-end;
-  padding: 10px;
-  margin-top: auto;
-`;
-
 const CloseButton = styled.button`
   position: absolute;
   width: 70px;
@@ -73,53 +66,6 @@ const CloseButton = styled.button`
   background-color: rgba(20, 46, 56, 0.9);
   color: white;
   border-radius: 0px 20px 0px 0px;
-`;
-
-const CartButton = styled.button`
-  position: absolute;
-  width: 115px;
-  height: 75px;
-  bottom: 0px;
-  right: 0px;
-  padding: 10px;
-  border: none;
-  background-color: transparent;
-  border: 1px solid var(--light-black);
-  border-radius: 0px 0px 20px 0px;
-  background-color: rgba(20, 46, 56, 0.9);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
-
-const CartImage = styled.img`
-  width: 30px;
-  height: 30px;
-  margin-bottom: 8px;
-`;
-
-const CartButtonText = styled.span`
-  color: var(--white);
-  font-size: 6px;
-`;
-
-const ImageBox = styled.div`
-  position: absolute;
-  background-color: #fab65d;
-  width: 155px;
-  height: 75px;
-  bottom: 0px;
-  right: 115px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 20px 0px 0px 0px;
-`;
-
-const Image = styled.img`
-  width: 50px;
-  height: 50px;
 `;
 
 const CustomContainer = styled.div`
@@ -155,16 +101,7 @@ const ModalComponent: React.FC<ModalProps> = ({ isOpen, onRequestClose, contentL
           <CustomSidebar />
           <CustomContent />
         </CustomContainer>
-        <ModalButtons>
-          <CloseButton onClick={onRequestClose}>X</CloseButton>
-          <ImageBox>
-            <Image src={modal_cake} alt="Cart" />
-          </ImageBox>
-          <CartButton>
-            <CartImage src={modal_cart} alt="Cart" />
-            <CartButtonText>장바구니 담기</CartButtonText>
-          </CartButton>
-        </ModalButtons>
+        <ModalButtons onRequestClose={onRequestClose} />
       </StyledModal>
     </ModalContainer>
   );
